+++ backend/src/main/java/com/nuriwoolim/pagebackend/domain/Post.java
@@ -1,10 +1,12 @@
-package com.nuriwoolim.pagebackend.entity;
+package com.nuriwoolim.pagebackend.domain;
 
 import com.nuriwoolim.pagebackend.core.BaseEntity;
-import com.nuriwoolim.pagebackend.user.entity.User;
+import com.nuriwoolim.pagebackend.domain.user.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.ConstraintMode;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ForeignKey;
 import jakarta.persistence.GeneratedValue;
@@ -12,6 +14,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
@@ -28,12 +31,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(
         indexes = {
-                @Index(name = "idx_schedule_user", columnList = "user_id"),
+                @Index(name = "idx_post_writer", columnList = "writer_id"),
+                @Index(name = "idx_post_board", columnList = "board_id")
         }
 )
 @Builder
 @AllArgsConstructor
-public class WeekSchedule extends BaseEntity {
+public class Post extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -41,17 +45,29 @@ public class WeekSchedule extends BaseEntity {
     @Column(unique = true, length = 20, nullable = false)
     private String title;
 
-    @Column(length = 100)
-    private String description;
+    @Column(length = 65_535, nullable = false)
+    @Lob
+    private String content;
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    @Builder.Default
+    private PostType type = PostType.GENERAL;
 
     @ManyToOne(fetch = jakarta.persistence.FetchType.LAZY, optional = false)
     @JoinColumn(
             nullable = false,
             foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
-    private User user;
+    private User writer;
 
-    @OneToMany(mappedBy = "weekSchedule", fetch = FetchType.LAZY, orphanRemoval = true,
+    @ManyToOne(fetch = jakarta.persistence.FetchType.LAZY, optional = false)
+    @JoinColumn(
+            nullable = false,
+            foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
+    private Board board;
+
+    @OneToMany(mappedBy = "post", fetch = FetchType.LAZY, orphanRemoval = true,
             cascade = {jakarta.persistence.CascadeType.REMOVE})
     @Builder.Default
-    private List<Schedule> scheduleList = new ArrayList<>();
+    private List<Comment> commentList = new ArrayList<>();
 }
