+++ backend/src/main/java/com/nuriwoolim/pagebackend/domain/user/entity/UserType.java
@@ -1,4 +1,4 @@
-package com.nuriwoolim.pagebackend.user.entity;
+package com.nuriwoolim.pagebackend.domain.user.entity;
 
 public enum UserType {
     ADMIN, MANAGER, MEMBER, NONMEMBER
