+++ backend/src/main/java/com/nuriwoolim/pagebackend/domain/user/service/UserService.java
@@ -1,11 +1,11 @@
-package com.nuriwoolim.pagebackend.user.service;
+package com.nuriwoolim.pagebackend.domain.user.service;
 
+import com.nuriwoolim.pagebackend.domain.user.dto.UserCreateRequest;
+import com.nuriwoolim.pagebackend.domain.user.dto.UserUpdateRequest;
+import com.nuriwoolim.pagebackend.domain.user.entity.User;
+import com.nuriwoolim.pagebackend.domain.user.repository.UserRepository;
 import com.nuriwoolim.pagebackend.global.exception.CustomException;
 import com.nuriwoolim.pagebackend.global.exception.ErrorCode;
-import com.nuriwoolim.pagebackend.user.dto.UserCreateRequest;
-import com.nuriwoolim.pagebackend.user.dto.UserUpdateRequest;
-import com.nuriwoolim.pagebackend.user.entity.User;
-import com.nuriwoolim.pagebackend.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
